# Conflicts:
#	pom.xml
#	src/main/java/com/avitam/fantasy11/model/User.java

+++ src/main/java/com/avitam/fantasy11/model/AddressRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressRepository extends MongoRepository<Address, Integer> {
+public interface AddressRepository extends MongoRepository<Address, Long> {
 }
