+++ src/main/java/com/avitam/fantasy11/model/PlayerRoleRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayerRoleRepository extends MongoRepository<PlayerRole, Integer> {
+public interface PlayerRoleRepository extends MongoRepository<PlayerRole, Long> {
 }
